+++ packages/atomic/src/components/facets-v1/facet-value-box/facet-value-box.tsx
@@ -1,9 +1,12 @@
 import {FunctionalComponent, h} from '@stencil/core';
 import {FacetValueProps} from '../facet-common';
 
-export const FacetValueBox: FunctionalComponent<FacetValueProps> = (props) => {
+export const FacetValueBox: FunctionalComponent<FacetValueProps> = (
+  props,
+  children
+) => {
   const count = props.numberOfResults.toLocaleString(props.i18n.language);
-  const ariaLabel = props.i18n.t('facetValue', {
+  const ariaLabel = props.i18n.t('facet-value', {
     value: props.displayValue,
     count: props.numberOfResults,
   });
@@ -19,15 +22,7 @@ export const FacetValueBox: FunctionalComponent<FacetValueProps> = (props) => {
         aria-pressed={props.isSelected.toString()}
         aria-label={ariaLabel}
       >
-        <span
-          title={props.displayValue}
-          part="value-label"
-          class={`value-label ellipsed w-full ${
-            props.isSelected ? 'font-bold' : ''
-          }`}
-        >
-          {props.displayValue}
-        </span>
+        {children}
         <span
           title={count}
           part="value-count"