+++ packages/atomic/src/components/facets-v1/facet-value-link/facet-value-link.tsx
@@ -1,9 +1,12 @@
 import {FunctionalComponent, h} from '@stencil/core';
 import {FacetValueProps} from '../facet-common';
 
-export const FacetValueLink: FunctionalComponent<FacetValueProps> = (props) => {
+export const FacetValueLink: FunctionalComponent<FacetValueProps> = (
+  props,
+  children
+) => {
   const count = props.numberOfResults.toLocaleString(props.i18n.language);
-  const ariaLabel = props.i18n.t('facetValue', {
+  const ariaLabel = props.i18n.t('facet-value', {
     value: props.displayValue,
     count: props.numberOfResults,
   });
@@ -17,13 +20,7 @@ export const FacetValueLink: FunctionalComponent<FacetValueProps> = (props) => {
         aria-pressed={props.isSelected.toString()}
         aria-label={ariaLabel}
       >
-        <span
-          title={props.displayValue}
-          part="value-label"
-          class={`value-label ellipsed ${props.isSelected ? 'font-bold' : ''}`}
-        >
-          {props.displayValue}
-        </span>
+        {children}
         <span
           part="value-count"
           class="ml-1.5 text-neutral-dark with-parentheses"