+++ packages/headless/doc-parser/src/headless-export-resolvers/initializer-resolver.ts
@@ -8,6 +8,16 @@ import {findApi} from '../api-finder';
 import {resolveFunction} from '../function-resolver';
 
 export function resolveInitializer(entry: ApiEntryPoint, name: string) {
+  const fn = findInitializer(entry, name);
+  return resolveFunction(entry, fn, [0]);
+}
+
+export function resolveEngineInitializer(entry: ApiEntryPoint, name: string) {
+  const fn = findInitializer(entry, name);
+  return resolveFunction(entry, fn, []);
+}
+
+function findInitializer(entry: ApiEntryPoint, name: string) {
   const fn = findApi(entry, name);
 
   if (!isFunction(fn)) {
@@ -16,7 +26,7 @@ export function resolveInitializer(entry: ApiEntryPoint, name: string) {
     );
   }
 
-  return resolveFunction(entry, fn, [0]);
+  return fn;
 }
 
 function isFunction(item: ApiItem): item is ApiFunction {