+++ packages/headless/src/features/question-answering/question-answering-actions-loader.ts
@@ -39,7 +39,7 @@ export interface QuestionAnsweringActionCreators {
 }
 
 /**
- * Loads the `query` reducer and returns possible action creators.
+ * Loads the `questionAnswering` reducer and returns possible action creators.
  *
  * @param engine - The headless engine.
  * @returns An object holding the action creators.