+++ packages/headless/src/product-recommendation.index.ts
@@ -1,4 +1,4 @@
-export {Unsubscribe} from '@reduxjs/toolkit';
+export {Unsubscribe, Middleware} from '@reduxjs/toolkit';
 
 export {
   ProductRecommendationEngine,
@@ -8,6 +8,15 @@ export {
   getSampleProductRecommendationEngineConfiguration,
 } from './app/product-recommendation-engine/product-recommendation-engine';
 
+export {CoreEngine, ExternalEngineOptions} from './app/engine';
+export {
+  EngineConfiguration,
+  AnalyticsConfiguration,
+  AnalyticsRuntimeEnvironment,
+} from './app/engine-configuration';
+export {LoggerOptions} from './app/logger';
+export {LogLevel} from './app/logger';
+
 // Actions
 export * from './features/configuration/configuration-actions-loader';
 export * from './features/product-recommendations/product-recommendations-actions-loader';
@@ -15,6 +24,11 @@ export * from './features/context/context-actions-loader';
 export * from './features/search-hub/search-hub-actions-loader';
 
 // Controllers
+export {
+  Controller,
+  buildController,
+} from './controllers/controller/headless-controller';
+
 export {
   FrequentlyBoughtTogetherListOptions,
   FrequentlyBoughtTogetherListProps,