+++ packages/quantic/force-app/main/default/lwc/quanticDateFacetValue/quanticDateFacetValue.js
@@ -1,8 +1,11 @@
 import {LightningElement, api} from 'lwc';
+import LOCALE from '@salesforce/i18n/locale';
 
 export default class QuanticDateFacetValue extends LightningElement {
   /** @type {import("coveo").DateFacetValue} */
   @api item;
+  /** @type {boolean} */
+  @api isChecked;
 
   /** @type {string} */
   @api start;
@@ -19,8 +22,8 @@ export default class QuanticDateFacetValue extends LightningElement {
 
   connectedCallback() {
     // eslint-disable-next-line  @lwc/lwc/no-api-reassignments
-    this.start = new Date(this.item.start).toLocaleDateString();
+    this.start = new Intl.DateTimeFormat(LOCALE).format(new Date(this.item.start));
     // eslint-disable-next-line  @lwc/lwc/no-api-reassignments
-    this.end = new Date(this.item.end).toLocaleDateString();
+    this.end = new Intl.DateTimeFormat(LOCALE).format(new Date(this.item.end));
   }
 }