+++ packages/quantic/force-app/main/default/lwc/quanticFacet/quanticFacet.js
@@ -1,5 +1,18 @@
-import {LightningElement, track, api} from 'lwc';
-import {registerComponentForInit, initializeWithHeadless} from 'c/quanticHeadlessLoader';
+import {
+  LightningElement,
+  track,
+  api
+} from 'lwc';
+import {
+  registerComponentForInit,
+  initializeWithHeadless
+} from 'c/quanticHeadlessLoader';
+import {I18nUtils} from 'c/quanticUtils';
+
+import showMore from '@salesforce/label/c.quantic_ShowMore';
+import showLess from '@salesforce/label/c.quantic_ShowLess';
+import showMoreFacetValues from '@salesforce/label/c.quantic_ShowMoreFacetValues';
+import showLessFacetValues from '@salesforce/label/c.quantic_ShowLessFacetValues';
 
 export default class QuanticFacet extends LightningElement {
   /** @type {import("coveo").FacetState} */
@@ -19,6 +32,13 @@ export default class QuanticFacet extends LightningElement {
   facet;
   /** @type {import("coveo").Unsubscribe} */
   unsubscribe;
+
+  labels = {
+    showMore,
+    showLess,
+    showMoreFacetValues,
+    showLessFacetValues,
+  }
 
   connectedCallback() {
     registerComponentForInit(this, this.engineId);
@@ -52,7 +72,10 @@ export default class QuanticFacet extends LightningElement {
   }
 
   get values() {
-    return this.state.values || [];
+    return this.state.values.map(v => ({
+      ...v,
+      checked: v.state === 'selected'
+    })) || [];
   }
 
   get canShowMore() {
@@ -71,6 +94,14 @@ export default class QuanticFacet extends LightningElement {
 
   get hasValues() {
     return this.values.length !== 0;
+  }
+
+  get showMoreFacetValuesLabel() {
+    return I18nUtils.format(this.labels.showMoreFacetValues, this.label)
+  }
+
+  get showLessFacetValuesLabel() {
+    return I18nUtils.format(this.labels.showLessFacetValues, this.label)
   }
 
   /**